Update Respiratory Decompensation & Failure.md
added "or adventitious" to breath sounds
I fixed question #3. because increased respirations could be interpreted as increased RR (which is correct)
Also, because of a sympathetic response, patients might experience hyperglycemia when in respiratory failure, so I changed this wrong answer as well. Replacement wrong answers: metabolic alkalosis and hypoglycemia

Depending on time available (in a perfect world) I would ask much higher level questions here-- eg. questions where a patient with specific symptoms presents and the nurse has to determine what is going on based on that assessment.