Update pheumothorax.md
This module only covers pneumothorax.  Was the expectation that chest tube management, pneumonia, and PEs were also to be covered?  I think at some point it would be good to add another module on chest tube management as this is a difficult one for nurses who may not see them regularly. However, for COVID, this is not critical, and could wait until after we launch these initial modules.