Monday
I narrowed the focus to the type of sedation used in critical care. Much of the info was based on the use of sedation in outpatient surgical procedures, so I removed those aspects. 
Need to change the image to the Modified ramsay scale to one that shows 8 stages. The scale showing now is the old Ramsay scale. I will try to send you one.
Remove image ![](assets/sedation4.png) as it is not pertinent
I have to teach for the next several hours so I can't add MCQs right now. I can try to add or edit some late this afternoon..